commit de terca

+++ src/Model/IMateriais.java
@@ -3,7 +3,7 @@ package Model;
 public interface IMateriais {
     public void setQuantidade(int quantidade);
     public void setProdutoNovo(String produtoNovo);
-    public void setPrecoNovo(int precoNovo);
+    public void setPrecoNovo(float precoNovo);
     public void setCodigoNovo(int codigoNovo);
     public void setDescricaoNovo(String descricaoNovo);
 }
